requirements.txt
streamlit==1.53.0
pandas==2.3.3
numpy==2.4.1
plotly==6.0.0
scipy==1.17.0
requests==2.32.5
python-dotenv==1.2.1
